% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_curve.R
\name{epidemic_curve}
\alias{epidemic_curve}
\title{Plot an epidemic curve with both bars and a trend line}
\usage{
epidemic_curve(data, date_col, count_col)
}
\arguments{
\item{data}{A data frame containing the date and case count columns.}

\item{date_col}{The name of the column with date information.}

\item{count_col}{The name of the column with case counts.}
}
\value{
A ggplot object representing the epidemic curve with a line.
}
\description{
Plot an epidemic curve with both bars and a trend line
}
