% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_ratio.R
\name{risk_ratio}
\alias{risk_ratio}
\title{Risk Ratio Plot}
\arguments{
\item{data}{A data frame containing exposure and outcome columns.}

\item{exposure_col}{The name of the column for exposure groups.}

\item{outcome_col}{The name of the column for outcomes.}
}
\value{
A ggplot object representing risk ratios.
}
\description{
This function computes and visualizes risk ratios for exposure and outcome groups.
}
